* refactor widgets into separate files * add threaded control_module to coordinator.control_module from sandbox example * eg. values.py
