+++ vent/gui/widgets/__init__.py
@@ -1,0 +1,4 @@
+from vent.gui.widgets.control import Control
+from vent.gui.widgets.monitor_value import Monitor_Value
+from vent.gui.widgets.plot import Plot
+from vent.gui.widgets.status_bar import Status_Bar
